ref: Allow printing of propagation config and remove templates (#718)
Remove the templating on the configuration structures and set the scalar types to float. Also provides printouts for the configuration. Adds a centralized options parsing to the detray tools, so that the boost options code does not need to be repeated in new executables. For this, the track generator configs had to be moved out of their parent classes to avoid templating them. The options parsing is calling add_options (adds the options to boost description), configure_options (checks the options and puts them in the configs) and print_options for every configuration that is given to it. All of these functions exist as template specializations, so that they can be forward declared.

Also fixes a bug in the file handle, where it would not respect the parent path when writing files without replacing them.

The changed floating point precision in the propagation config seem to have made the wire chamber test fail in one track each, so I tweaked the parameters a little

+++ tests/integration_tests/device/cuda/telescope_navigation_validation.cpp
@@ -67,8 +67,10 @@ int main(int argc, char **argv) {
     cfg_str_nav.name("telescope_detector_straight_line_navigation_cuda");
     cfg_str_nav.whiteboard(white_board);
     auto mask_tolerance = cfg_ray_scan.mask_tolerance();
-    cfg_str_nav.propagation().navigation.min_mask_tolerance = mask_tolerance[0];
-    cfg_str_nav.propagation().navigation.max_mask_tolerance = mask_tolerance[1];
+    cfg_str_nav.propagation().navigation.min_mask_tolerance =
+        static_cast<float>(mask_tolerance[0]);
+    cfg_str_nav.propagation().navigation.max_mask_tolerance =
+        static_cast<float>(mask_tolerance[1]);
 
     detail::register_checks<detray::cuda::straight_line_navigation>(
         tel_det, tel_names, cfg_str_nav);
